+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Fix format of the default value for required `Options` in the help.
+- Match command by partial name, if it is unique.
 
 ## 1.0.0 - 2019-08-01
 - Initial implementation